+++ src/dlls/engine/10_expgfx/expgfx.c
@@ -566,9 +566,8 @@ void objfx_spawnDirectionalBurst(void* obj, u8 idx, f32 scale, u8 kind, u8 mode,
 
 #define OBJ_FX_PI 3.1415927f
 
-void objfx_spawnArcedBurst(void* obj, u8 idx, f32 scale, u8 kind, u8 mode, int chance, f32 angBase, f32 lo, f32 hi,
-                           void* origin, int flags)
-{
+void objfx_spawnArcedBurst(void* obj, u8 idx, f32 scale, u8 kind, u8 mode, u8 chance, f32 angBase, f32 lo, f32 hi,
+                           void* origin, int flags) {
     ObjFxParticleParams params;
     ObjFxU16Table9 effectParams = *(ObjFxU16Table9*)((char*)&gObjFxCrystalSparkleTbl + 0x8c);
     ObjFxU16Table8 spawnIds = *(ObjFxU16Table8*)((char*)&gObjFxCrystalSparkleTbl + 0xa0);
@@ -583,12 +582,10 @@ void objfx_spawnArcedBurst(void* obj, u8 idx, f32 scale, u8 kind, u8 mode, int c
     params.scale = scale;
     params.effectParam = effectParams.values[(u8)kind];
     params.pad00[1] = 0x3c;
-    for (i = 0; i < 4; i++)
-    {
+    for (i = 0; i < 4; i++) {
         u16 val;
         f32 a;
-        if (randomGetRange(0, 0x63) >= (u8)chance)
-        {
+        if (randomGetRange(0, 0x63) >= chance) {
             continue;
         }
         rvec[0] = randomGetRange(0, 0xffff);
@@ -598,8 +595,7 @@ void objfx_spawnArcedBurst(void* obj, u8 idx, f32 scale, u8 kind, u8 mode, int c
         angularT = randomGetRange(0, 1000) / 1000.0f;
         params.position[1] = 0.0f;
         params.position[2] = 0.0f;
-        switch ((u8)mode)
-        {
+        switch (mode) {
         case 1:
             params.position[0] = 1.0f - radialT * radialT;
             break;
@@ -637,8 +633,7 @@ void objfx_spawnArcedBurst(void* obj, u8 idx, f32 scale, u8 kind, u8 mode, int c
             f32 t = angularT - 0.5f;
             params.position[1] = t * hi;
         }
-        if (origin != NULL)
-        {
+        if (origin != NULL) {
             params.position[0] += ((GameObject*)origin)->anim.localPosX;
             params.position[1] += ((GameObject*)origin)->anim.localPosY;
             params.position[2] += ((GameObject*)origin)->anim.localPosZ;
